Accept ecdsa- prefix in defaultKeyCommand output
It was already possible to use ecdsa-* signing keys, but this did not
work with the ssh.ggp.defaultKeyCommand because the output was only
checked for keys starting with "ssh-".

Introduces a new function valid_ssh_signing_key that is used wherever
the code previously checked ` starts_with(key, "ssh-")`.